+++ src/main/java/duke/command/AddDeadlineCommand.java
@@ -1,12 +1,12 @@
-package command;
+package duke.command;
 
-import tasks.Deadline;
-import tasks.TaskList;
-import util.Ui;
-import util.Storage;
+import duke.tasks.Deadline;
+import duke.tasks.TaskList;
+import duke.util.Ui;
+import duke.util.Storage;
 
 /**
- * Represents a command to add a deadline task to the task list.
+ * Represents a duke.command to add a deadline task to the task list.
  * <p>
  * The {@code AddDeadlineCommand} class encapsulates the details of the deadline,
  * including its description and due date. When executed, the deadline is added
@@ -38,10 +38,10 @@ public class AddDeadlineCommand extends Command {
     }
 
     /**
-     * Executes the command by adding a new deadline to the provided task list,
+     * Executes the duke.command by adding a new deadline to the provided task list,
      * and saving the updated list using the provided storage.
      *
-     * @param tasks list of tasks.
+     * @param tasks list of duke.tasks.
      * @param ui user interface.
      * @param storage storage system.
      */
@@ -52,9 +52,9 @@ public class AddDeadlineCommand extends Command {
     }
 
     /**
-     * Indicates that this command doesn't terminate the application.
+     * Indicates that this duke.command doesn't terminate the application.
      *
-     * @return {@code false} since this command doesn't cause the application to exit.
+     * @return {@code false} since this duke.command doesn't cause the application to exit.
      */
     @Override
     public boolean isExit() {
