+++ src/main/java/duke/command/UnmarkCommand.java
@@ -1,14 +1,14 @@
-package command;
+package duke.command;
 
-import exception.InvalidIndexException;
-import tasks.TaskList;
-import util.Ui;
-import util.Storage;
+import duke.exception.InvalidIndexException;
+import duke.tasks.TaskList;
+import duke.util.Ui;
+import duke.util.Storage;
 
 /**
- * Represents a command to unmark a task as completed.
+ * Represents a duke.command to unmark a task as completed.
  * <p>
- * When executed, this command will unmark the task at the specified index as completed.
+ * When executed, this duke.command will unmark the task at the specified index as completed.
  * </p>
  */
 public class UnmarkCommand extends Command {
@@ -26,9 +26,9 @@ public class UnmarkCommand extends Command {
     }
 
     /**
-     * Executes the unmark command, unmarking the task at the specified index as completed.
+     * Executes the unmark duke.command, unmarking the task at the specified index as completed.
      *
-     * @param tasks List of tasks.
+     * @param tasks List of duke.tasks.
      * @param ui User interface.
      * @param storage Storage system.
      * @throws InvalidIndexException If the given index is out of bounds.
