+++ src/main/java/duke/exception/EmptyTaskException.java
@@ -1,6 +1,6 @@
-package exception;
+package duke.exception;
 
-import tasks.TaskType;
+import duke.tasks.TaskType;
 
 /**
  * Exception thrown when a task's field (description, by, from, to) is found to be empty.
@@ -8,7 +8,7 @@ import tasks.TaskType;
 public class EmptyTaskException extends Exception{
 
     /**
-     * Constructs an exception.EmptyTaskException with a specific error message based on the type of task and the empty field.
+     * Constructs an duke.exception.EmptyTaskException with a specific error message based on the type of task and the empty field.
      *
      * @param type  The type of task (e.g., todo, deadline, event).
      * @param field The specific field (e.g., description, by, from, to) that is empty.
