+++ src/main/java/duke/exception/InvalidDateTimeException.java
@@ -1,9 +1,9 @@
-package exception;
+package duke.exception;
 
 /**
- * Represents an exception that is thrown when the application encounters an invalid date and time input.
+ * Represents an duke.exception that is thrown when the application encounters an invalid date and time input.
  * <p>
- * This exception is triggered when the user provides a date and time format that does not match the expected formats:
+ * This duke.exception is triggered when the user provides a date and time format that does not match the expected formats:
  * "yyyy-MM-dd" or "yyyy-MM-dd HHmm".
  * </p>
  */
