+++ src/main/java/duke/tasks/TaskType.java
@@ -1,0 +1,8 @@
+package duke.tasks;
+
+/**
+ * Enum representing the types of duke.tasks.
+ */
+public enum TaskType {
+    TODO, DEADLINE, EVENT
+}
