+++ src/main/java/duke/util/Pair.java
@@ -1,4 +1,4 @@
-package util;
+package duke.util;
 
 /**
  * Represents an immutable key-value pair.
