--- src/main/java/tasks/TaskType.java
@@ -1,8 +1,0 @@
-package tasks;
-
-/**
- * Enum representing the types of tasks.
- */
-public enum TaskType {
-    TODO, DEADLINE, EVENT
-}
