+++ text-ui-test/EXPECTED.TXT
@@ -5,20 +5,20 @@ What can I do for you?
 ----------------------------------------
 Got it. I've added this task:
 	[T][ ] borrow book
-Now you have 1 tasks in the list.
+Now you have 1 duke.tasks in the list.
 ----------------------------------------
 ----------------------------------------
 Got it. I've added this task:
 	[D][ ] return book (by: Dec 11 2019, 6:00pm)
-Now you have 2 tasks in the list.
+Now you have 2 duke.tasks in the list.
 ----------------------------------------
 ----------------------------------------
 Got it. I've added this task:
 	[E][ ] project meeting (from: Dec 11 2019, 6:00pm to: Jan 21 2020, 9:30pm)
-Now you have 3 tasks in the list.
+Now you have 3 duke.tasks in the list.
 ----------------------------------------
 ----------------------------------------
-Here are the tasks in your list:
+Here are the duke.tasks in your list:
 1.[T][ ] borrow book
 2.[D][ ] return book (by: Dec 11 2019, 6:00pm)
 3.[E][ ] project meeting (from: Dec 11 2019, 6:00pm to: Jan 21 2020, 9:30pm)
@@ -28,7 +28,7 @@ Nice! I've marked this task as done:
 	[T][X] borrow book
 ----------------------------------------
 ----------------------------------------
-Here are the tasks in your list:
+Here are the duke.tasks in your list:
 1.[T][X] borrow book
 2.[D][ ] return book (by: Dec 11 2019, 6:00pm)
 3.[E][ ] project meeting (from: Dec 11 2019, 6:00pm to: Jan 21 2020, 9:30pm)
@@ -40,10 +40,10 @@ OK, I've marked this task as not done yet:
 ----------------------------------------
 Noted. I've removed this task:
 	[E][ ] project meeting (from: Dec 11 2019, 6:00pm to: Jan 21 2020, 9:30pm)
-Now you have 2 tasks in the list.
+Now you have 2 duke.tasks in the list.
 ----------------------------------------
 ----------------------------------------
-Here are the tasks in your list:
+Here are the duke.tasks in your list:
 1.[T][ ] borrow book
 2.[D][ ] return book (by: Dec 11 2019, 6:00pm)
 ----------------------------------------
